- Make Document.DegreesRotated a short, since it really can only take 4 values (0, 90, 180, 270). It is now a short (signed) in both the application and the database;
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@674 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/TableTypes/Document.cs
@@ -137,10 +137,10 @@ namespace OpenDentBusiness{
 		}
 		///<summary>Only allowed 0,90,180, and 270.</summary>
 		[DataField("DegreesRotated")]
-		private int degreesRotated;
+		private short degreesRotated;
 
 		/// <summary>Only allowed 0,90,180, and 270.</summary>
-		public int DegreesRotated {
+		public short DegreesRotated {
 			get { return degreesRotated; }
 			set { degreesRotated = value; MarkDirty(); degreesRotatedChanged = true; }
 		}
